+++ agents/mongodb/src/version/v3_0/models.rs
@@ -4,7 +4,7 @@ use serde_derive::Deserialize;
 use replicante_agent::Result;
 use replicante_models_agent::info::ShardRole;
 
-use super::super::super::error::ErrorKind;
+use crate::error::ErrorKind;
 
 /// Section of the replSetGetStatus command that we care about.
 #[derive(Debug, Deserialize)]
@@ -82,7 +82,6 @@ impl ReplSetStatusMember {
 
 #[cfg(test)]
 mod tests {
-    use bson::bson;
     use bson::doc;
     use bson::Bson;
     use lazy_static::lazy_static;