+++ agents/mongodb/src/version/v3_0/replica.rs
@@ -1,13 +1,9 @@
 use std::sync::Arc;
 
-use bson::bson;
 use bson::doc;
 use bson::Bson;
 use failure::ResultExt;
-use mongodb::db::ThreadedDatabase;
 use mongodb::Client;
-use mongodb::CommandType;
-use mongodb::ThreadedClient;
 use opentracingrust::utils::FailSpan;
 use opentracingrust::Log;
 use opentracingrust::Span;
@@ -58,8 +54,8 @@ impl ReplicaSet {
             .start_timer();
         let info = self
             .client
-            .db("test")
-            .command(doc! {"buildInfo" => 1}, CommandType::BuildInfo, None)
+            .database("test")
+            .run_command(doc! {"buildInfo": 1}, None)
             .fail_span(&mut span)
             .map_err(|error| {
                 MONGODB_OP_ERRORS_COUNT
@@ -88,8 +84,8 @@ impl ReplicaSet {
             .start_timer();
         let status = self
             .client
-            .db("admin")
-            .command(doc! {"replSetGetStatus" => 1}, CommandType::IsMaster, None)
+            .database("admin")
+            .run_command(doc! {"replSetGetStatus" => 1}, None)
             .fail_span(&mut span)
             .map_err(|error| {
                 MONGODB_OP_ERRORS_COUNT