+++ libs/rust/sdk/src/api/agent/info.rs
@@ -1,5 +1,6 @@
 use std::sync::Arc;
 
+use actix_web::web::Data;
 use actix_web::HttpRequest;
 use actix_web::HttpResponse;
 use actix_web::Responder;
@@ -12,13 +13,11 @@ use replicante_util_tracing::fail_span;
 use crate::Agent;
 
 /// API interface to Agent::agent_info
-pub fn agent(request: HttpRequest) -> Result<impl Responder> {
+pub async fn agent(request: HttpRequest, agent: Data<Arc<dyn Agent>>) -> Result<impl Responder> {
     let mut exts = request.extensions_mut();
     let mut span = request_span(&mut exts);
     span.log(Log::new().log("span.kind", "server-receive"));
-    let info = request
-        .app_data::<Arc<dyn Agent>>()
-        .expect("dyn Agent must be available as App::data")
+    let info = agent
         .agent_info(&mut span)
         .map_err(|error| fail_span(error, &mut *span))?;
     let response = HttpResponse::Ok().json(info);
@@ -27,43 +26,26 @@ pub fn agent(request: HttpRequest) -> Result<impl Responder> {
 }
 
 /// API interface to Agent::datastore_info
-pub struct DatastoreInfo {
-    agent: Arc<dyn Agent>,
-    cluster_display_name_override: Option<String>,
-}
+pub async fn datastore(
+    request: HttpRequest,
+    agent: Data<Arc<dyn Agent>>,
+    cluster_display_name_override: Data<Option<String>>,
+) -> Result<impl Responder> {
+    let mut exts = request.extensions_mut();
+    let mut span = request_span(&mut exts);
+    span.log(Log::new().log("span.kind", "server-receive"));
+    let mut info = agent
+        .datastore_info(&mut span)
+        .map_err(|error| fail_span(error, &mut *span))?;
 
-impl DatastoreInfo {
-    pub fn new(
-        agent: Arc<dyn Agent>,
-        cluster_display_name_override: Option<String>,
-    ) -> DatastoreInfo {
-        DatastoreInfo {
-            agent,
-            cluster_display_name_override,
-        }
-    }
-}
+    // Inject the cluster_display_name override if configured.
+    info.cluster_display_name = cluster_display_name_override
+        .as_ref()
+        .as_ref()
+        .cloned()
+        .or(info.cluster_display_name);
 
-impl Responder for DatastoreInfo {
-    type Error = actix_web::Error;
-    type Future = Result<HttpResponse, actix_web::Error>;
-
-    fn respond_to(self, request: &HttpRequest) -> Self::Future {
-        let mut exts = request.extensions_mut();
-        let mut span = request_span(&mut exts);
-        span.log(Log::new().log("span.kind", "server-receive"));
-        let mut info = self
-            .agent
-            .datastore_info(&mut span)
-            .map_err(|error| fail_span(error, &mut *span))?;
-
-        // Inject the cluster_display_name override if configured.
-        info.cluster_display_name = self
-            .cluster_display_name_override
-            .or(info.cluster_display_name);
-
-        let response = HttpResponse::Ok().json(info);
-        span.log(Log::new().log("span.kind", "server-send"));
-        Ok(response)
-    }
+    let response = HttpResponse::Ok().json(info);
+    span.log(Log::new().log("span.kind", "server-send"));
+    Ok(response)
 }