+++ libs/rust/sdk/src/api/agent/mod.rs
@@ -9,24 +9,13 @@ use replicante_util_actixweb::TracingMiddleware;
 pub mod info;
 pub mod shards;
 
-use self::info::DatastoreInfo;
-use self::shards::Shards;
-
 use super::APIRoot;
-use crate::Agent;
 use crate::AgentContext;
 
 /// Configure all agent endpoints.
-pub fn configure_app(
-    flags: &APIFlags,
-    app: &mut web::ServiceConfig,
-    agent: Arc<dyn Agent>,
-    context: &AgentContext,
-) {
+pub fn configure_app(flags: &APIFlags, app: &mut web::ServiceConfig, context: &AgentContext) {
     let tracer = Arc::clone(&context.tracer);
     APIRoot::UnstableAPI.and_then(flags, |root| {
-        let agent_for_datastore = Arc::clone(&agent);
-        let agent_for_shards = agent;
         let cluster_display_name_override = context.config.cluster_display_name_override.clone();
         app.service(
             root.resource("/info/agent")
@@ -38,16 +27,12 @@ pub fn configure_app(
         );
         app.service(
             root.resource("/info/datastore")
+                .data(cluster_display_name_override)
                 .wrap(TracingMiddleware::new(
                     context.logger.clone(),
                     Arc::clone(&tracer),
                 ))
-                .route(web::get().to(move || {
-                    DatastoreInfo::new(
-                        Arc::clone(&agent_for_datastore),
-                        cluster_display_name_override.clone(),
-                    )
-                })),
+                .route(web::get().to(info::datastore)),
         );
         app.service(
             root.resource("/shards")
@@ -55,7 +40,7 @@ pub fn configure_app(
                     context.logger.clone(),
                     Arc::clone(&tracer),
                 ))
-                .route(web::get().to(move || Shards::new(Arc::clone(&agent_for_shards)))),
+                .route(web::get().to(shards::shards)),
         );
     });
 }