+++ libs/rust/sdk/src/api/index.rs
@@ -1,7 +1,7 @@
 use actix_web::HttpResponse;
 use actix_web::Responder;
 
-pub fn index() -> impl Responder {
+pub async fn index() -> impl Responder {
     HttpResponse::Ok().body("Replicante Agent API endpoints")
 }
 
@@ -15,13 +15,14 @@ mod tests {
     use actix_web::web;
     use actix_web::App;
 
-    #[test]
-    fn index_points_to_api() {
-        let mut app = init_service(App::new().route("/", web::get().to(super::index)));
+    #[actix_rt::test]
+    async fn index_points_to_api() {
+        let app = init_service(App::new().route("/", web::get().to(super::index)));
+        let mut app = app.await;
         let request = TestRequest::default().to_request();
-        let response = call_service(&mut app, request);
+        let response = call_service(&mut app, request).await;
         assert_eq!(response.status(), StatusCode::OK);
-        let body = read_body(response);
+        let body = read_body(response).await;
         assert_eq!(
             String::from_utf8(body.to_vec()).unwrap(),
             "Replicante Agent API endpoints"