+++ libs/rust/sdk/src/api/introspect/threads.rs
@@ -6,7 +6,7 @@ use humthreads::registered_threads;
 use humthreads::ThreadStatus;
 
 /// Expose a snaphot view of traked threads states.
-pub fn handler() -> impl Responder {
+pub async fn handler() -> impl Responder {
     let mut threads = registered_threads();
     threads.sort_unstable_by_key(|t| t.name.clone());
     let threads = ThreadsResponse::new(threads);